CREATE DECISION-MAKING
THE FOLLOWING CODE CAN TELL WHETHER A VOTEE IS ELIGIBLE TO VOTE OR NOT:
(1)  FOR THE PERSON WHO IS NOT ELIGIBLE: it can further tell that after how many years he/she could give vote.
(2) FOR THE PERSON WHO IS ELIGIBLE: the code further asks the user to enter its gender. After entering, it then greets the user and then asks them to choose between the political parties between A,B,C, and if the user enters any other letter apart from the following. the vote will be given to none party.